chore: Improved behavior tree class and skill structure
