+++ Assets/Common/Constants.cs
@@ -1,5 +1,5 @@
 public class Constants
 {
     public const float Play_TIME = 60 * 3;
-    public const float Respawn_TIME = 5;
+    public const float RESPAWN_TIME = 5;
 }
