+++ Assets/Scenes/Battle/Summons/Script/BasicClass/BehaviourTree.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f593842db2e97634c8569a623e55a331
+guid: 492a16518f18c484fa492c2a4959d3d6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
