+++ Assets/Scenes/Battle/Summons/Script/BasicClass/Skills.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 01f74fad31a94d5439e5e4e8a0ebcef6
+guid: e614c2346d271e741ad6e23e185a8c51
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
