--- Assets/Scenes/Battle/Summons/Script/BehaviorTree/Inverter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dd0bca0922f3d8c4eb106839e22855ea
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
