--- Assets/Scenes/Battle/Summons/Script/BehaviorTree/Parallel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d526eec311f818648988a31621cb6a8e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
