--- Assets/Scenes/Battle/Summons/Script/BehaviorTree/Selector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: effd9414eafd2624788529be5a547d83
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
