--- Assets/Scenes/Battle/Summons/Script/BehaviorTree/Tree.cs
@@ -1,20 +1,0 @@
-using UnityEngine;
-
-namespace BehaviorTree
-{
-    public abstract class Tree : MonoBehaviour
-    {
-        private Node _root = null;
-
-        protected void Start()
-        {
-            _root = SetupTree();
-        }
-        private void Update()
-        {
-            if (_root != null)
-                _root.Evaluate();
-        }
-        protected abstract Node SetupTree();
-    }
-}
