--- Assets/Scenes/Battle/Summons/Script/BehaviorTree/Tree.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fdd2ccea68dafaf468d9db32278379fc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
