+++ Assets/Scenes/Battle/Summons/Script/SenorZorro.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 using BehaviorTree;
+using Skills;
 
 public class SenorZorro : Summon
 {
