+++ Assets/Scenes/Battle/Summons/Script/Summon.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 using BehaviorTree;
+using Skills;
 
 //공통적이되 함수 내용이 달라지는 클래스
 public abstract class Summon : MonoBehaviour
@@ -108,59 +109,18 @@ public abstract class Summon : MonoBehaviour
             case Enums.ECC.Stun:
                 break;
             case Enums.ECC.KnockBack:
+                //StartCoroutine(KnockBack());
                 break;
             default:
                 break;
         }
     }
+    //private IEnumerator KnockBack()
+    //{
+       // yield return mWait;
+       // Vector3 playerPos = GameManager.instance.Player.transform.position;
+       // Vector3 dirVec = transform.position - playerPos;
+       // mRb.AddForce(dirVec.normalized * (3 + mKnockbackpower), ForceMode2D.Impulse);
+    // }
     #endregion
-}
-
-public class Skill
-{
-    public float skiilCounter = 0;
-    protected float[] stats;
-    protected float skillCooldown = 0f;
-
-    public virtual void Execute(GameObject summon, GameObject target, Animator animator)
-    {
-        // 동작 구현
-    }
-    public float[] Stats
-    {
-        get { return stats; }
-        set { stats = value; }
-    }
-    #region 쿨타임
-    public bool IsCooldown()
-    {
-        return skillCooldown > 0f;
-    }
-    public void StartCooldown()
-    {
-        skillCooldown = stats[(int)Enums.ESkillStats.CoolTime];
-    }
-    public void UpdateCooldown(float deltaTime)
-    {
-        if (stats[((int)Enums.ESkillStats.CoolTime)] > 0f)
-        {
-            skillCooldown -= deltaTime;
-            if (skillCooldown <= 0f)
-            {
-                skillCooldown = 0f;
-            }
-        }
-    }
-    #endregion
-    public void FlipSprite(GameObject summon, GameObject target)
-    {
-        if (summon.transform.position.x < target.transform.position.x)
-        {
-            summon.GetComponent<SpriteRenderer>().flipX = true;
-        }
-        else
-        {
-            summon.GetComponent<SpriteRenderer>().flipX = false;
-        }
-    }
 }
